+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changes from v0.4 to current version:
 * Added document describing the operations being carried out by the lensing engine when it draws
   shears according to a user-specified power spectrum. (Issue #248)
 
+* Added the ability to draw lensing shears and convergences self-consistently
+  from the same input shear power spectrum.  (Issue #304)
+
+* Added a utility that can take an input set of shears on a grid, and
+  reconstruct the convergence.  (Issue #304)
+
 * Added Shapelet class (sub-class of GSObject) for describing shapelet profiles. (Issue #350)
 
 * Made various speed improvements related to drawing images, both in real and Fourier space.